+++ examples/unix/00_hello/00_hello.c
@@ -11,7 +11,7 @@
  ************************************************************************/
 #define TICOS_HEARTBEAT_MILLIS 1000
 
-#include <ticos.h>
+#include <ticos_framework.h>
 #include <ticos/log.h>
 
 /************************************************************************
