+++ library.properties
@@ -7,5 +7,5 @@ paragraph=This library provides the support of Ticos IoT Framework using in Ardu
 category=Other
 url=https://github.com/tiwater/ticos-framework
 architectures=*
-includes=ticos.h
+includes=ticos_framework.h
 depends=ticos-sdk-for-c
