+++ platforms/unix/inc/ticos/platform.h
