+++ scripts/cmake/ticos.cmake
@@ -6,5 +6,5 @@ aux_source_directory (${TICOS_PLATFORM_DIR}/src  PLATFORM_SRCS)
 if (TICOS_USE_FREERTOS)
 aux_source_directory (${TICOS_SOURCE_DIR}/os/freertos OS_SRCS)
 endif ()
-set (FULL_SRCS ${PLATFORM_SRCS} ${OS_SRCS} ${CORE_SRCS})
+set (FRAMEWORK_SRCS ${PLATFORM_SRCS} ${OS_SRCS} ${CORE_SRCS})
 set (TICOS_INCS ${TICOS_PLATFORM_DIR}/inc ${TICOS_SOURCE_DIR})
