+++ scripts/codegen/templates/main_c
@@ -11,7 +11,7 @@
  ************************************************************************/
 #define TICOS_HEARTBEAT_MILLIS 250
 
-#include <ticos.h>
+#include <ticos_framework.h>
 
 /************************************************************************
  * 向工程注册总线标识号，请将此标识号列表成员作为TICOS_BUS_ADD首参
