+++ scripts/codegen/templates/proj_tmpl_esp32/cmake_lists_main
@@ -1,6 +1,6 @@
 idf_component_register (
         SRCS
-        ${FULL_SRCS}
+        ${FRAMEWORK_SRCS}
         ${PROJ_SRCS}
         ${TICOS_SRCS}
         INCLUDE_DIRS
