+++ scripts/codegen/templates/ticos_proj_tmpl/esp32/main/CMakeLists.txt
@@ -1,6 +1,6 @@
 idf_component_register (
         SRCS
-        ${FULL_SRCS}
+        ${FRAMEWORK_SRCS}
         ${TICOS_SRCS}
         ${PROJ_SRCS}
         INCLUDE_DIRS
